+++ example/lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_dropdown/flutter_dropdown.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_dropdown/flutter_dropdown.dart';
 
 void main() => runApp(MyApp());
 
@@ -60,36 +60,12 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
     return Scaffold(
       appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title),
       ),
       body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
         child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             Text(
@@ -101,6 +77,35 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
             DropDown<String>(
               items: <String>["Male", "Female", "Other"],
+              customWidgets: <Widget>[
+                Row(
+                  children: <Widget>[
+                    Text("Male"),
+                    CircleAvatar(
+                      backgroundImage: NetworkImage(
+                          "https://raw.githubusercontent.com/rrousselGit/provider/master/resources/expanded_devtools.jpg"),
+                    ),
+                  ],
+                ),
+                Row(
+                  children: <Widget>[
+                    Text("Female"),
+                    CircleAvatar(
+                      backgroundImage: NetworkImage(
+                          "https://raw.githubusercontent.com/rrousselGit/provider/master/resources/expanded_devtools.jpg"),
+                    ),
+                  ],
+                ),
+                Row(
+                  children: <Widget>[
+                    Text("Other"),
+                    CircleAvatar(
+                      backgroundImage: NetworkImage(
+                          "https://raw.githubusercontent.com/rrousselGit/provider/master/resources/expanded_devtools.jpg"),
+                    ),
+                  ],
+                )
+              ],
               hint: "Select gender",
               onChange: print,
             ),